# openapi.json
{"openapi":"3.1.0","info":{"title":"RAG API — Lecture 17","description":"Upload files → Ask questions → Get answers with exact source references","version":"1.0.0"},"paths":{"/upload":{"post":{"summary":"Upload Files","description":"Upload one or many files.\nEach file is loaded, split into chunks, embedded, and stored in Qdrant.\nReturns a per-file status — one failure does not cancel the rest.","operationId":"upload_files_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_files_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UploadResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/query":{"post":{"summary":"Query","description":"Ask a question — returns the answer and every source chunk used to build it.","operationId":"query_query_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QueryRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QueryResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/documents":{"get":{"summary":"List Documents","description":"List all files currently indexed in Qdrant (one row per file, not per chunk).","operationId":"list_documents_documents_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocumentListResponse"}}}}}}},"delete":{"delete":{"summary":"Delete Document","description":"Remove a file and ALL its chunks from Qdrant by doc_id.","operationId":"delete_documentdelete_delete","parameters":[{"name":"doc_id","in":"query","required":true,"schema":{"type":"string","title":"Doc Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health","description":"Quick check that the server is running and shows the active configuration.","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Health Health Get"}}}}}}}},"components":{"schemas":{"Body_upload_files_upload_post":{"properties":{"files":{"items":{"type":"string","contentMediaType":"application/octet-stream"},"type":"array","title":"Files","description":"One or more files: PDF, DOCX, MD, CSV, TXT"}},"type":"object","required":["files"],"title":"Body_upload_files_upload_post"},"DeleteResponse":{"properties":{"status":{"type":"string","title":"Status"},"doc_id":{"type":"string","title":"Doc Id"},"file_name":{"type":"string","title":"File Name"},"chunks_deleted":{"type":"integer","title":"Chunks Deleted"}},"type":"object","required":["status","doc_id","file_name","chunks_deleted"],"title":"DeleteResponse","description":"Confirmation that a document was removed."},"DocumentInfo":{"properties":{"doc_id":{"type":"string","title":"Doc Id"},"file_name":{"type":"string","title":"File Name"},"file_type":{"type":"string","title":"File Type"},"chunk_count":{"type":"integer","title":"Chunk Count"},"upload_timestamp":{"type":"string","title":"Upload Timestamp"}},"type":"object","required":["doc_id","file_name","file_type","chunk_count","upload_timestamp"],"title":"DocumentInfo","description":"One row in the list of indexed documents."},"DocumentListResponse":{"properties":{"documents":{"items":{"$ref":"#/components/schemas/DocumentInfo"},"type":"array","title":"Documents"},"total":{"type":"integer","title":"Total"}},"type":"object","required":["documents","total"],"title":"DocumentListResponse","description":"Response for GET /documents."},"FileUploadResult":{"properties":{"file_name":{"type":"string","title":"File Name"},"doc_id":{"type":"string","title":"Doc Id"},"file_type":{"type":"string","title":"File Type"},"status":{"type":"string","title":"Status"},"chunks_created":{"type":"integer","title":"Chunks Created"},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error"}},"type":"object","required":["file_name","doc_id","file_type","status","chunks_created"],"title":"FileUploadResult","description":"Status for a single file in a batch upload."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"QueryRequest":{"properties":{"question":{"type":"string","title":"Question"},"top_k":{"type":"integer","maximum":20.0,"minimum":1.0,"title":"Top K","default":5}},"type":"object","required":["question"],"title":"QueryRequest","description":"What the user sends when asking a question."},"QueryResponse":{"properties":{"question":{"type":"string","title":"Question"},"answer":{"type":"string","title":"Answer"},"references":{"items":{"$ref":"#/components/schemas/Reference"},"type":"array","title":"References"},"embedding_provider":{"type":"string","title":"Embedding Provider"}},"type":"object","required":["question","answer","references","embedding_provider"],"title":"QueryResponse","description":"What we send back: the answer plus every source used to generate it."},"Reference":{"properties":{"file_name":{"type":"string","title":"File Name"},"file_type":{"type":"string","title":"File Type"},"page_number":{"type":"integer","title":"Page Number"},"chunk_index":{"type":"integer","title":"Chunk Index"},"chunk_text":{"type":"string","title":"Chunk Text"},"relevance_score":{"type":"number","title":"Relevance Score"}},"type":"object","required":["file_name","file_type","page_number","chunk_index","chunk_text","relevance_score"],"title":"Reference","description":"One chunk that was retrieved and used to build the answer."},"UploadResponse":{"properties":{"total_files":{"type":"integer","title":"Total Files"},"successful":{"type":"integer","title":"Successful"},"failed":{"type":"integer","title":"Failed"},"results":{"items":{"$ref":"#/components/schemas/FileUploadResult"},"type":"array","title":"Results"}},"type":"object","required":["total_files","successful","failed","results"],"title":"UploadResponse","description":"Summary of a batch upload — one FileUploadResult per file."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}